requirements.txt
# -i https://pypi.org/simple
anyio==3.6.1
click==8.1.3
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.85.0
h11==0.13.0
idna==3.4
passlib==1.7.4
pyasn1==0.4.8
pydantic[email]==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlalchemy==1.4.41
starlette==0.20.4
typing-extensions==4.3.0
uvicorn==0.18.3
